analysis/requirements.txt
pandas
numpy
matplotlib
openpyxl
pyyaml
